front: using types generated by rtk
